+++ cmd/goreleaser/internal/configure.go
@@ -29,11 +29,13 @@ import (
 )
 
 const (
-	armArch          = "arm"
-	coreDistro       = "otelcol"
-	contribDistro    = "otelcol-contrib"
-	k8sDistro        = "otelcol-k8s"
-	otlpDistro       = "otelcol-otlp"
+	armArch            = "arm"
+	coreDistro         = "otelcol"
+	contribDistro      = "otelcol-contrib"
+	k8sDistro          = "otelcol-k8s"
+	otlpDistro         = "otelcol-otlp"
+	ebpfProfilerDistro = "otelcol-ebpf-profiler"
+	dockerHub        = "otel"
 	ghcr             = "ghcr.io/open-telemetry/opentelemetry-collector-releases"
 	binaryNamePrefix = "otelcol"
 	imageNamePrefix  = "opentelemetry-collector"
@@ -45,8 +47,9 @@ var (
 	winContainerArchs = []string{"amd64"}
 	darwinArchs       = []string{"amd64", "arm64"}
 	k8sArchs          = []string{"amd64", "arm64", "ppc64le", "s390x"}
+	ebpfProfilerArchs = []string{"amd64"}
 
-	imageRepos = []string{ghcr}
+	imageRepos = []string{dockerHub, ghcr}
 
 	// otelcol (core) distro
 	otelColDist = newDistributionBuilder(coreDistro).WithConfigFunc(func(d *distribution) {
@@ -140,6 +143,22 @@ var (
 		d.containerImageManifests = slices.Concat(
 			newContainerImageManifests(d.name, "linux", k8sArchs, containerImageOptions{}),
 		)
+	}).WithDefaultArchives().WithDefaultChecksum().WithDefaultSigns().WithDefaultDockerSigns().WithDefaultSBOMs().Build()
+
+	// ebpf-profiler distro
+	ebpfProfilerDist = newDistributionBuilder(ebpfProfilerDistro).WithConfigFunc(func(d *distribution) {
+		d.buildConfigs = []buildConfig{
+			&fullBuildConfig{targetOS: "linux", targetArch: ebpfProfilerArchs},
+		}
+		d.containerImages = slices.Concat(
+			newContainerImages(d.name, "linux", ebpfProfilerArchs, containerImageOptions{}),
+		)
+		d.containerImageManifests = slices.Concat(
+			newContainerImageManifests(d.name, "linux", ebpfProfilerArchs, containerImageOptions{}),
+		)
+		d.enableCgo = true
+		d.ldFlags = "-extldflags=-static"
+		d.goTags = "osusergo,netgo"
 	}).WithDefaultArchives().WithDefaultChecksum().WithDefaultSigns().WithDefaultDockerSigns().WithDefaultSBOMs().Build()
 )
 
@@ -610,6 +629,8 @@ func BuildDist(dist string, onlyBuild bool) config.Project {
 		return otlpDist.BuildProject()
 	case k8sDistro:
 		return k8sDist.BuildProject()
+	case ebpfProfilerDistro:
+		return ebpfProfilerDist.BuildProject()
 	case contribDistro:
 		if onlyBuild {
 			return contribBuildOnlyDist.BuildProject()